Added support for new motion correction methods in meso
MotionCorrection now supports multiple methods of correction. The structure is as follows:

shared.MotionCorrectionMethod + meso.MotionMethodForScan -> meso.MotionCorrection -> meso.MotionCorrection.MotionMethodUsed

Method 8 is currently suggested as the best method and is automatically filled out in MotionMethodForScan once RasterCorrection is done populating. This ensures users do not need to perform extra steps to process their scans. The method used is also recorded in a new part table, MotionMethodUsed. This is done because MotionCorrection cannot have MotionMethodForScan added as a dependency without recreating the table, so keys in MotionMethodForScan could be changed/deleted without it propagating downstream.