Fixing critical timing bug in clocktools
Found a critical bug that caused errors when fetching timing/interpolation data for "fluorescence-*" or "deconvolution-*" settings.

When including a `unit_id` or `mask_id` in the key for convert_clocks, the internal function `fetch_timing_data` would pull from the ms_delay stored in `pipe.ScanSet.UnitInfo` and add it to the current timestamps. However, timestamps are stored in seconds while `ms_delay` was converted to milliseconds. `ms_delay` was previously not divided by 1,000 before being added to the current timestamps, leading to reported times that can be 1s - 20s off. This would also make units that are fairly synchronized appear completely uncorrelated.

When including just `field` in the key for convert_clocks, the field delay was calculated from the delay image stored in `pipe.ScanInfo.Field`. There was a bug where this was using the average value of the delay image, but delay images are based off of the first pixel recorded at the top depth. That means when clocktools later sliced the timestamps unique for the field's depth, extra delay was added. This delay was on the order of milliseconds (20ms - 200ms), so it should be less consequential than the unit_id issue mentioned above.